src/docker/django-docker/djangoapp/requirements.txt
aioredis==1.3.1
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-rest==0.0.1
django-rest-auth==0.9.5
django-rest-registration==0.5.4
djangorestframework==3.11.0
drf-yasg==1.17.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msgpack==0.6.2
mysqlclient==1.4.6
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
Twisted==19.10.0
txaio==20.1.1
typing==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.8
zope.interface==4.7.2
haikunator
jsonfield
